fix(BR-206) Adicionado a mensagem de confirmação de cancelar criação de grupo em todas as paginas
