+++ src/app/(tabs)/groups/newGroup/index.tsx
@@ -21,15 +21,12 @@ import useKeyboardStatus from '@/src/utils/keyboardUtils'
 import { axiosClient } from '@/src/utils/axios'
 import { verticalScale } from '@/src/utils/responsiveUtils'
 
-import { ButtonCustomizer } from '@/src/components/ButtonCustomizer'
-
 import Photograph from '@/src/assets/images/photograph.svg'
 import Pencil from '@/src/assets/images/pencil.svg'
 
 import { theme } from '@/src/theme'
-import { styles as globalStyles } from '@/src/app/styles'
 import { styles } from './styles'
-import ModalComponent from '@/src/components/ModalComponent'
+import Footer from '../components/Footer'
 
 interface FormData {
   name: string
@@ -52,9 +49,8 @@ export default function NewGroup() {
   const [groupImage, setGroupImage] = useState<string>('')
   const [categoriesList, setCategoriesList] = useState<Categories[]>([])
   const [selectedCategory, setSelectedCategory] = useState<string>('')
-  const [modalVisible, setModalVisible] = useState(false)
 
-  const { groupData, setGroupData, removeGroupData } = useGroupStore()
+  const { groupData, setGroupData } = useGroupStore()
 
   const {
     control,
@@ -81,11 +77,6 @@ export default function NewGroup() {
     if (!result.canceled) {
       setGroupImage(result.assets[0].base64 ?? '')
     }
-  }
-
-  const handleCancelAddGroup = () => {
-    removeGroupData()
-    router.replace('/groups')
   }
 
   useEffect(() => {
@@ -133,15 +124,6 @@ export default function NewGroup() {
 
   return (
     <View style={styles.container}>
-      <ModalComponent
-        type="complete"
-        title="Deseja realmente sair?"
-        description="Você perderá a edição do grupo"
-        textButton1="Cancelar"
-        textButton2="Confirmar"
-        onPress={() => handleCancelAddGroup()}
-        showModal={[modalVisible, setModalVisible]}
-      />
       <TouchableWithoutFeedback onPress={Keyboard.dismiss}>
         <View style={{ flex: 1, gap: 8 }}>
           {/* Header */}
@@ -250,31 +232,7 @@ export default function NewGroup() {
       </TouchableWithoutFeedback>
 
       {/* Botão */}
-      {!isKeyboardVisible && (
-        <View style={{ flexDirection: 'row', justifyContent: 'space-between' }}>
-          <ButtonCustomizer.Root
-            type="tertiaryHalfWidth"
-            onPress={() => {
-              setModalVisible(true)
-            }}
-          >
-            <ButtonCustomizer.Title
-              title="Cancelar"
-              customStyles={globalStyles.secondaryButtonText}
-            />
-          </ButtonCustomizer.Root>
-
-          <ButtonCustomizer.Root
-            type="primaryHalfWidth"
-            onPress={handleSubmit(handleData)}
-          >
-            <ButtonCustomizer.Title
-              title="Proximo"
-              customStyles={globalStyles.primaryButtonText}
-            />
-          </ButtonCustomizer.Root>
-        </View>
-      )}
+      {!isKeyboardVisible && <Footer handleSubmit={handleSubmit(handleData)} />}
     </View>
   )
 }
